// react-frontend/src/api/models/order-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FeedbackDTO } from './feedback-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { MenuItemDTO } from './menu-item-dto';

/**
 * 
 * @export
 * @interface OrderItemDTO
 */
export interface OrderItemDTO {
    /**
     * 
     * @type {number}
     * @memberof OrderItemDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof OrderItemDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof OrderItemDTO
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof OrderItemDTO
     */
    'price'?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderItemDTO
     */
    'quantity'?: number;
    /**
     * 
     * @type {string}
     * @memberof OrderItemDTO
     */
    'deliveryDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof OrderItemDTO
     */
    'totalPrice'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof OrderItemDTO
     */
    'allergies'?: Array<string>;
    /**
     * 
     * @type {MenuItemDTO}
     * @memberof OrderItemDTO
     */
    'menuItem'?: MenuItemDTO;
    /**
     * 
     * @type {FeedbackDTO}
     * @memberof OrderItemDTO
     */
    'feedback'?: FeedbackDTO;
    /**
     * 
     * @type {string}
     * @memberof OrderItemDTO
     */
    'status'?: StatusEnum;
    /**
     * 
     * @type {string}
     * @memberof OrderItemDTO
     */
    'editBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof OrderItemDTO
     */
    'dateTimeDelivered'?: string;
}

export const StatusEnum = {
    PAID: 'PAID',
    IN_PROGRESS: 'IN_PROGRESS',
    IN_DELIVERY: 'IN_DELIVERY',
    DELIVERED: 'DELIVERED',
    CANCELLED: 'CANCELLED',
    REFUNDED: 'REFUNDED'
} as const;

export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];